Tilemap viewer

+++ Core/Debugger/PpuTools.cpp
@@ -41,7 +41,7 @@ uint8_t PpuTools::GetTilePixelColor(const uint8_t* ram, const uint32_t ramMask, 
 
 void PpuTools::BlendColors(uint8_t output[4], uint8_t input[4])
 {
-	uint8_t alpha = input[3] + 1;
+	int alpha = input[3] + 1;
 	uint8_t invertedAlpha = 256 - input[3];
 	output[0] = (uint8_t)((alpha * input[0] + invertedAlpha * output[0]) >> 8);
 	output[1] = (uint8_t)((alpha * input[1] + invertedAlpha * output[1]) >> 8);
